refactor(core, test): extract `TMXLSResourceResolver` and improve TMX validation in tests
- Extracted `TMXLSResourceResolver` for reusable schema and DTD resolution.
- Updated `TMXWriterTest` to include schema validation using `tmx14.xsd`.
- Removed redundant `DumbLSResourceResolver` from `TMXReader2`.
